feat(ui): add label formatting and paddings
